Refactoring
Split off the memory access code into a "Memory" class.
Rename a bunch of things to be more clear such as the menu items.
Move the startup dialog to be accessed via About... in the context menu.
Use enums for the settings options.
Stop using arrays for memory addresses, use selects and constants.