Added check for updates functionality, alert and update reason.
Uses the version number stored in TPO_latest.txt on my Dropbox as the definitive latest version and update reason/patch notes.

From now on, releases will involve incrementing the Settings.VERSION number string, AND updating the TPO_latest.txt file to reflect the latest version and the reason for the update.